create tournament.py and rename Game to Pong
