+++ django/src/games/routing.py
@@ -1,7 +1,8 @@
 from django.urls import re_path
-from . import pong
+from pong import Pong
+from tournament import Tournament
 
 websocket_urlpatterns = [
-    re_path(r'ws/pong/?$', pong.Game.as_asgi()),
-    re_path(r'ws/pong_tournament/?$', pong.Tournament.as_asgi()),
+    re_path(r'ws/pong/?$', Pong.as_asgi()),
+    re_path(r'ws/pong_tournament/?$', Tournament.as_asgi()),
 ]
